Update unit-project-1-starter-code.ipynb
Admission is a yes or no which is of type categorical. 
Predictors/covariates are everything but admission. Admission is the outcome. 
There is a timeframe, but it's unknown. 
Missing admission problem statement. 
EDA, yes. Also whether there are other things to discover different from the hypothesis, whether data is missing or needs to be transformed.
Outliers skew all of the data in the direction of the outlier.